Add TensorFlow format support to tf.keras.Model.save_weights and load_weights
Supports restore-on-create in subclassed Models when executing eagerly, and removes the requirement that the Model be built before weights are loaded.

Currently only subclassed Models work with the TensorFlow weight format. Graph networks will need a bit of extra logic to support the same topology/by-name distinction as the current HDF5 format (and for now they don't even add Checkpointable dependencies on their sub-layers).

Some notes:
  - Checkpoints won't be numbered. This keeps behavior the same as for existing HDF5 weight saving.
  - All dependencies will be saved for subclassed Models, not just layers. This will make it more useful for training checkpoints (you can assign an optimizer to an attribute and save the slot variables that way).
  - Subclassed models won't support loading by flattened weight list from the TensorFlow format. Since there's no global naming for Layers (it's local to the Model), I think this is reasonable.

PiperOrigin-RevId: 193965120